requirements.txt
openai>=1.30.0
customtkinter==5.2.2
tkinterdnd2==0.3.0
pydub==0.25.1
fpdf2==2.7.9
openpyxl==3.1.2
pillow>=10.0.0
pyinstaller>=6.15.0
